requirements.txt
python-vlc
termcolor